201 Learning Path Page
Refactored TopicContentView to break it out some into another component
to handle adding a button to collapse the list of events for the topic
view on mobile which I also added in this commit. I used semantic
accordion to handle this, but may want to use another method. Semantic
doesn't seem to have a better way.