308 Modified learning path dropdown for logged in state
Modified the learning path dropdown to not show the objective
sub menus if the user is not logged in. When logged in the calls
to get objectives are masqueraded, and if you are not enrolled it
will break the nav links. Could be handled better with additional
checks if objectives for enrolled courses are desired.