Dashboard conditionals
Added conditionals to check if a User is logged in or not and what they
will see.

If a user is logged in, <DashboardHeader /> and <DashboardCTA /> will
show.

If a user is not registered or not logged in, then <DashboardSplash />
will show.

We will need to add a a connection to the Redux store to check the state
of a user being logged in or not.