Added action and reducers to check last LP
Added an action to get the latest Learning Path accessed. If a user has
a last Learning Path it will show the LearningPathResume component.

If the user does not have a last Learning Path it will show call to
action.

If the user is not logged in it will show the intro modal.