SBAELIDEV-357 Created Actions and Reducers
I created actions and reducers to check if the user has started a
journey. If the user has started a journey, then they will see the
LearningPathsResume component.

I also created an action to get the latest enrollment which will be used
to resume a journey. This passes in the course id to LearningPathResume
so it can fetch the learning path/journey, get the progress, and update
the component with the information of completed MLEs.