Additional logic fix for learning path component update
Gave more thought to how this component was updating. Fixed error in
how I was declaring variables, and additionally adjusted the logic more.
This component should only need to fire off the initialFunctions if the
path id changes. Calling off the event id and topic id would be incorrect
and cause more calls to be made than needed.